+++ gradle-plugin/src/main/groovy/mozilla/telemetry/glean-gradle-plugin/GleanGradlePlugin.groovy
@@ -535,7 +535,7 @@ except:
     void apply(Project project) {
         isOffline = project.gradle.startParameter.offline
 
-        project.ext.glean_version = "66.3.0"
+        project.ext.glean_version = "67.0.0"
         def parserVersion = gleanParserVersion(project)
 
         // Print the required glean_parser version to the console. This is